7skeleton-master/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-cors==0.7.0
aiokafka==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.8.2
contourpy==1.3.0
croniter==2.0.7
cycler==0.12.1
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.0
faust-streaming==0.11.3
folium==0.18.0
fonttools==4.54.1
frozenlist==1.4.1
GeoAlchemy2==0.15.2
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
intervaltree==3.1.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
mapclassify==2.8.1
MarkupSafe==3.0.2
matplotlib==3.9.2
mode-streaming==0.4.1
motor==3.6.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.2
opentracing==2.4.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
psycopg-binary==3.2.3
pyarrow==17.0.0
pydantic==1.10.18
pydantic_core==2.23.4
pymongo==4.9.2
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.38.6
terminaltables==3.1.10
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
venusian==3.1.0
watchfiles==0.24.0
websockets==13.1
xyzservices==2024.9.0
yarl==1.14.0
